tests: manual changes
